COVID-19 Geovisualisation Documentation

COVID-19 Geovisualisation Documentation

Geovisualisation.jl

utils.jl

Index

Geovisualisation.jl

Geovisualisation.runGeovisualiser — Method

runGeovisualiser()

utils.jl

Geovisualisation.readGlobalConfirmedCSV — Method

readGlobalConfirmedCSV()

Reads time series COVID-19 global confirmed cases raw CSV file from repository.

Returns dataframe of global confirmed cases.

Geovisualisation.readGlobalDeathsCSV — Method

readGlobalDeathsCSV()

Reads time series COVID-19 global deaths raw CSV file from repository.

Returns dataframe of global death cases.

Geovisualisation.readGlobalRecoveredCSV — Method

readGlobalRecoveredCSV

Reads time series COVID-19 global recovered cases raw CSV file from repository.

Returns dataframe of global recovered cases.

Geovisualisation.readGlobalVaccinationCSV — Method

readGlobalVaccinationCSV

Reads time series COVID-19 global recovered cases raw CSV file from repository.

Returns dataframe of global vaccination cases.

Geovisualisation.getTotalConfirmedCases — Method

getTotalConfirmedCases

Gets total confirmed cases worldwide by summing all countries in dataframe up to specified date.

Returns number of total confirmed cases of country at a specific date.

Geovisualisation.getTotalRecoveredCases — Method

getTotalRecoveredCases

Gets total recovered cases worldwide by summing all countries in dataframe up to specified date.

Returns number of total recovered cases of country at a specific date.

Geovisualisation.getTotalDeaths — Method

getTotalDeathsCases

Gets total deaths cases worldwide by summing all countries in dataframe up to specified date.

Returns number of total deaths cases of country at a specific date.

Geovisualisation.getTotalVaccinations — Method

getTotalVaccinations

Gets total vaccinations worldwide by summing all countries in dataframe.

Returns number of total vaccinations in a country at a specific date.

Geovisualisation.getTotalCaseFatality — Method

getTotalCaseFatality

Gets total case fatality rate worldwide.

Returns case fatality rate in a country at a specific date.

Geovisualisation.getCaseFatalityDataframe — Method

getCaseFatalityDataframe

Gets case fatality rate dataframe from confirmed and death dataframes.

Returns case fatality rate dataframe.

Geovisualisation.getListOfCountries — Method

getListOfCountries

Gets list of all countries listed in dataframe. If country has multiple entries, the province/state is added also.

Returns array of countries for dropdown.

Geovisualisation.getStartDate — Method

getStartDate

Provides start date for date picker.

Returns date object of start date.

Geovisualisation.getEndDate — Method

getEndDate

Provides end date for date picker.

Returns date object of end date.

Geovisualisation.formatToDateObject — Method

formatToDateObject

Converts date format provided in the dataframe to a date object.

Returns date object of date string.

Geovisualisation.convertTimeSeriesData — Method

convertTimeSeriesData

Converts array of integer timeseries data into regular values.

Returns array of values.

Index

Geovisualisation.convertTimeSeriesData

Geovisualisation.formatToDateObject

Geovisualisation.getCaseFatalityDataframe

Geovisualisation.getEndDate

Geovisualisation.getListOfCountries

Geovisualisation.getStartDate

Geovisualisation.getTotalCaseFatality

Geovisualisation.getTotalConfirmedCases

Geovisualisation.getTotalDeaths

Geovisualisation.getTotalRecoveredCases

Geovisualisation.getTotalVaccinations

Geovisualisation.readGlobalConfirmedCSV

Geovisualisation.readGlobalDeathsCSV

Geovisualisation.readGlobalRecoveredCSV

Geovisualisation.readGlobalVaccinationCSV

Geovisualisation.runGeovisualiser