COVID-19 Geovisualisation Documentation

Introduction

This web application is an interactive geographical visualiser based on COVID-19 data provided by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University.

The global view GUI consists of two sections, general information on the left and an interactive geographical visualisation scatter plot on the right. The information cards provide up to date information from the COVID-19 dataset. This information can be changed/updated using the controls in the navigation bar. Each information card has an additional tooltip. A tooltip provides additional information on the attribute. Also on the left hand-side is the top six table. The interactive geographical visualisation scatter plot provides an interactive visual aspect to the project. It displays data using a scatter plot graph. Only one attribute is displayed at any one time. The attributes can be changed using the tabs above the scatter map plot. The options are confirmed cases, death cases and case fatality rate. This also updates the data on table. The slider under the scatter map plot allows the user to select a date to view on the map plot. This ranges from the beginning of the COVID-19 pandemic to current day. A description on how to use the slider is located beside it.

The country view GUI consists of two sections, information cards and statistics on the left and interactive line graphs on the right. The information cards provides up to date information from the COVID-19 dataset. This information can be changed/updated using the controls in the navigation bar. Each information card has an additional tooltip also. The statistics table provides additional detailed information on the selected county, such as highest confirmed cases in a single day or even the mean confirmed cases. On the second section of the country view, it is split into four graphs. The first row consists of a map plot of the selected country and a line graph of vaccinations administered. The second row is made up of line graphs for confirmed cases and deaths. All of the graphs/plots are interactive and use the selected countries data from the datasets.

Table of Contents

COVID-19 Geovisualisation Documentation

Introduction

Table of Contents

Geovisualisation.jl

utils.jl

Index

Geovisualisation.jl

Geovisualisation.jl contain the Dash web application, including the layout and callbacks.

Geovisualisation.runGeovisualiser — Method

runGeovisualiser()

utils.jl

Geovisualisation.readGlobalConfirmedCSV — Method

readGlobalConfirmedCSV()

Reads time series COVID-19 global confirmed cases raw CSV file from repository.

Returns dataframe of global confirmed cases.

Geovisualisation.readGlobalDeathsCSV — Method

readGlobalDeathsCSV()

Reads time series COVID-19 global deaths raw CSV file from repository.

Returns dataframe of global death cases.

Geovisualisation.readGlobalRecoveredCSV — Method

readGlobalRecoveredCSV

Reads time series COVID-19 global recovered cases raw CSV file from repository.

Returns dataframe of global recovered cases.

Geovisualisation.readGlobalVaccinationCSV — Method

readGlobalVaccinationCSV

Reads time series COVID-19 global recovered cases raw CSV file from repository.

Returns dataframe of global vaccination cases.

Geovisualisation.getTotalConfirmedCases — Method

getTotalConfirmedCases

Gets total confirmed cases worldwide by summing all countries in dataframe up to specified date.

Returns number of total confirmed cases of country at a specific date.

Geovisualisation.getTotalRecoveredCases — Method

getTotalRecoveredCases

Gets total recovered cases worldwide by summing all countries in dataframe up to specified date.

Returns number of total recovered cases of country at a specific date.

Geovisualisation.getTotalDeaths — Method

getTotalDeathsCases

Gets total deaths cases worldwide by summing all countries in dataframe up to specified date.

Returns number of total deaths cases of country at a specific date.

Geovisualisation.getTotalVaccinations — Method

getTotalVaccinations

Gets total vaccinations worldwide by summing all countries in dataframe.

Returns number of total vaccinations in a country at a specific date.

Geovisualisation.getTotalCaseFatality — Method

getTotalCaseFatality

Gets total case fatality rate worldwide.

Returns case fatality rate in a country at a specific date.

Geovisualisation.getTopSixFromDataframe — Method

getTopSixFromDataframe

Gets top six countries for selected attribute worldwide.

Returns country and value for table.

Geovisualisation.getStatisticsTableForCountry — Method

getStatisticsTableForCountry

Gets statistics for table for specified country. Dataframes are already filtered.

Returns statistics with values for table in country view

Geovisualisation.getCaseFatalityDataframe — Method

getCaseFatalityDataframe

Gets case fatality rate dataframe from confirmed and death dataframes.

Returns case fatality rate dataframe.

Geovisualisation.getListOfCountries — Method

getListOfCountries

Gets list of all countries listed in dataframe. If country has multiple entries, the province/state is added also.

Returns array of countries for dropdown.

Geovisualisation.getStartDate — Method

getStartDate

Provides start date for date picker.

Returns date object of start date.

Geovisualisation.getEndDate — Method

getEndDate

Provides end date for date picker.

Returns date object of end date.

Geovisualisation.formatToDateObject — Method

formatToDateObject

Converts date format provided in the dataframe to a date object.

Returns date object of date string.

Geovisualisation.convertTimeSeriesData — Method

convertTimeSeriesData

Converts array of integer timeseries data into regular values.

Returns array of values.

Geovisualisation.prettifyNumberArray — Method

prettifyNumberArray

This function prettifies a number array to include comma separated values.

Returns array of prettified numbers.

Geovisualisation.roundNumberArray — Method

prettifyNumberArray

This function rounds a number array to 3 decimal places.

Returns array of round numbers.

Index

Geovisualisation.convertTimeSeriesData

Geovisualisation.formatToDateObject

Geovisualisation.getCaseFatalityDataframe

Geovisualisation.getEndDate

Geovisualisation.getListOfCountries

Geovisualisation.getStartDate

Geovisualisation.getStatisticsTableForCountry

Geovisualisation.getTopSixFromDataframe

Geovisualisation.getTotalCaseFatality

Geovisualisation.getTotalConfirmedCases

Geovisualisation.getTotalDeaths

Geovisualisation.getTotalRecoveredCases

Geovisualisation.getTotalVaccinations

Geovisualisation.prettifyNumberArray

Geovisualisation.readGlobalConfirmedCSV

Geovisualisation.readGlobalDeathsCSV

Geovisualisation.readGlobalRecoveredCSV

Geovisualisation.readGlobalVaccinationCSV

Geovisualisation.roundNumberArray

Geovisualisation.runGeovisualiser